secretstroll/part3/requirements_fingerprinting.txt
numpy
tqdm
scapy
scikit-learn
matplotlib
python-pcapng
dpkt
scipy
tensorflow-metal
tensorflow-macos
pandas